requirements-script.txt
pandas
cloudinary
requests
python-dotenv
google-api-python-client
